refactor: enum 데이터 패키징화에 따른 수정
