+++ src/main/java/com/windry/chordplayer/api/SongAPI.java
@@ -1,9 +1,8 @@
 package com.windry.chordplayer.api;
 
-import com.windry.chordplayer.domain.Gender;
-import com.windry.chordplayer.domain.SearchCriteria;
-import com.windry.chordplayer.domain.SortStrategy;
-import com.windry.chordplayer.domain.Tuning;
+import com.windry.chordplayer.spec.Gender;
+import com.windry.chordplayer.spec.SearchCriteria;
+import com.windry.chordplayer.spec.SortStrategy;
 import com.windry.chordplayer.dto.CreateSongDto;
 import com.windry.chordplayer.dto.FiltersOfSongList;
 import com.windry.chordplayer.dto.SongListItemDto;
@@ -11,6 +10,7 @@ import com.windry.chordplayer.exception.InvalidInputException;
 import com.windry.chordplayer.service.SongService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -29,7 +29,7 @@ public class SongAPI {
             @RequestBody CreateSongDto createSongDto
     ) {
         Long song = songService.createNewSong(createSongDto);
-        return ResponseEntity.ok().body(song);
+        return ResponseEntity.status(HttpStatus.CREATED).body(song);
     }
 
     @GetMapping("")
@@ -44,7 +44,7 @@ public class SongAPI {
             @RequestParam(value = "sort", required = false) SortStrategy sortStrategy
     ) {
 
-        if(page == null || size == null)
+        if (page == null || size == null)
             throw new InvalidInputException();
 
         FiltersOfSongList filters = FiltersOfSongList.builder()
