+++ src/main/java/com/windry/chordplayer/domain/Lyrics.java
@@ -1,5 +1,6 @@
 package com.windry.chordplayer.domain;
 
+import com.windry.chordplayer.spec.Tag;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
