+++ src/main/java/com/windry/chordplayer/spec/SearchCriteria.java
@@ -1,4 +1,4 @@
-package com.windry.chordplayer.domain;
+package com.windry.chordplayer.spec;
 
 public enum SearchCriteria {
     TITLE,
