+++ src/main/java/com/windry/chordplayer/spec/Tuning.java
@@ -1,4 +1,4 @@
-package com.windry.chordplayer.domain;
+package com.windry.chordplayer.spec;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
